feat!: label with source IP
For environments which have multiple stable egress IP addresses, it can
be helpful to expose this as part of the metrics so that we can see
which IP address is getting close to the allowed usage limits.

BREAKING CHANGE: `dockerhub_limit_remaining_requests_total` and
`dockerhub_limit_max_requests_total` now have a `source_ip` label.